Возвращает отрисовку точек маршрута и обработчики событий в Main
