+++ src/view/board/board.js
@@ -1,10 +1,5 @@
-import {IS_NEW_MODE} from "../../utils/constants";
-import {createElement, render} from "../../utils/render";
+import {createElement} from "../../utils/render";
 import {createBoard} from "./templates/create-board";
-import {getBlankPoint} from "../../temp/mocks";
-
-import EventView from "../event/event";
-import EventEditView from "../event-edit/event-edit";
 
 export default class Board {
   constructor(points) {
@@ -27,58 +22,5 @@ export default class Board {
   removeElement() {
     this._element = null;
   }
-
-  _renderEvent(boardElement, point) {
-    const eventElement = new EventView(point).getElement();
-    const eventEditElement = new EventEditView(point).getElement();
-
-    const onEscKeyDown = (evt) => {
-      if (evt.key === `Escape` || evt.key === `Esc`) {
-        evt.preventDefault();
-        setViewMode();
-        document.removeEventListener(`keydown`, onEscKeyDown);
-      }
-    };
-
-    const setEditMode = () => {
-      boardElement.replaceChild(eventEditElement, eventElement);
-    };
-
-    const setViewMode = () => {
-      boardElement.replaceChild(eventElement, eventEditElement);
-    };
-
-    eventElement.querySelector(`.event__rollup-btn`)
-      .addEventListener(`click`, () => {
-        setEditMode();
-        document.addEventListener(`keydown`, onEscKeyDown);
-      });
-
-    eventEditElement.querySelector(`.event__rollup-btn`)
-      .addEventListener(`click`, () => {
-        setViewMode();
-        document.removeEventListener(`keydown`, onEscKeyDown);
-      });
-
-    eventEditElement.querySelector(`.event__reset-btn`)
-      .addEventListener(`click`, () => {
-        setViewMode();
-        document.removeEventListener(`keydown`, onEscKeyDown);
-      });
-
-    eventEditElement.querySelector(`form`)
-      .addEventListener(`submit`, (evt) => {
-        evt.preventDefault();
-      });
-
-    render(boardElement, eventElement);
-  }
-
-  init() {
-    render(this.getElement(), new EventEditView(getBlankPoint(), IS_NEW_MODE).getElement());
-
-    this._points.forEach((point) => {
-      this._renderEvent(this.getElement(), point);
-    });
-  }
 }
+
